backend/requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
gunicorn
requests
APScheduler
Werkzeug
bcrypt
flask-limiter
